+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Fixed `ImportError` when using `isort>=5`
+  ( https://github.com/WanzenBug/pylint-import-requirements/issues/27 )
 
 ## [2.0.4] - 2020-08-03
 ### Fixed